authentication

+++ controllers/courses controller.js
@@ -46,10 +46,17 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 */
 exports.createCourse = asyncHandler(async (req, res, next) => {
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
     if (!bootcamp) {
-        return next(new ErrorResponse(`No bootcamp found with the id ${req.params.bootcampId}`));
+        return next(new ErrorResponse(`No bootcamp found with the id ${req.params.bootcampId}`, 404));
     }
+
+    // Make sure that the user logged in is the bootcamp owner
+    if (req.user.id !== bootcamp.user.toString() && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`user with id ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`));
+    }
+
     const course = await Course.create(req.body);
     res.status(200).json({
         success: true,
@@ -62,15 +69,23 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
     @access Private
 */
 exports.updateCourse = asyncHandler(async (req, res, next) => {
-    
-    const course = await Course.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true
-    })
+    let course = await Course.findById(req.params.id);
 
     if (!course) {
         return next(new ErrorResponse(`No course found with the id ${req.params.id}`));
     }
+    console.log(`here is the course ${course}`);
+
+    // Make sure user is course owner
+    if (req.user.id !== course.user.toString() && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`user with the id ${req.user.id} is not authorized to update this course ${course._id}`));
+    }
+
+    course = await Course.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+        runValidators: true
+    })
+
     res.status(200).json({
         success: true,
         data: course
@@ -88,6 +103,12 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
     if (!course) {
         return next(new ErrorResponse(`No course found with the id ${req.params.id}`));
     }
+    console.log(course.user)
+    // Make sure user is course owner
+    if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`user with the id ${req.user.id} is not authorized to delete this course ${course._id}`));
+    }
+    
     await course.deleteOne();
     res.status(200).json({
         success: true,
